# swagger.yml
openapi: 3.0.3
info:
  title: SEARCH Gateway Service
  description: SEARCH Gateway Service using Flask, OpenAPI and Connexion
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Rest Server for SEARCH-AI-Gateway
paths:
...
  /l3s-recsys/test-post:
    post:
      tags:
        - L3S Services
      operationId: l3s_gateway.recsysTestPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestDto'
        required: true
      responses:
        201:
          description: test post method for l3s recsys service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestDto'
      x-codegen-request-body-name: payload
  /l3s-recsys/random:
    post:
      tags:
        - L3S Services
      operationId: l3s_gateway.recsysRandom
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RandomRequest'
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RandomResponse'
        201:
          description: Success
          content: {}
        400:
          description: Number of recommendation is either negative or exceed the number
            of existing contents
          content: {}
      x-codegen-request-body-name: payload
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    TestDto:
      required:
      - message
      type: object
      properties:
        message:
          type: string
    RandomRequest:
      required:
        - num_of_rec
      type: object
      properties:
        num_of_rec:
          type: integer
    RandomResponse:
      required:
        - id
      type: object
      properties:
        id:
          type: string