# swagger_client_L3S/l3s_recsys_client/models/random_request.py
# coding: utf-8

"""
    L3S Recommendation Service for SEARCH

    Welcome to the Swagger UI documentation site!  # noqa: E501

    OpenAPI spec version: 0.0.1
    
    Generated by: https://github.com/swagger-api/swagger-codegen.git
"""

import pprint
import re  # noqa: F401

import six

class RandomRequest(object):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    """
    Attributes:
      swagger_types (dict): The key is attribute name
                            and the value is attribute type.
      attribute_map (dict): The key is attribute name
                            and the value is json key in definition.
    """
    swagger_types = {
        'num_of_rec': 'int'
    }

    attribute_map = {
        'num_of_rec': 'num_of_rec'
    }

    def __init__(self, num_of_rec=None):  # noqa: E501
        """RandomRequest - a model defined in Swagger"""  # noqa: E501
        self._num_of_rec = None
        self.discriminator = None
        self.num_of_rec = num_of_rec

    @property
    def num_of_rec(self):
        """Gets the num_of_rec of this RandomRequest.  # noqa: E501


        :return: The num_of_rec of this RandomRequest.  # noqa: E501
        :rtype: int
        """
        return self._num_of_rec

    @num_of_rec.setter
    def num_of_rec(self, num_of_rec):
        """Sets the num_of_rec of this RandomRequest.


        :param num_of_rec: The num_of_rec of this RandomRequest.  # noqa: E501
        :type: int
        """
        if num_of_rec is None:
            raise ValueError("Invalid value for `num_of_rec`, must not be `None`")  # noqa: E501

        self._num_of_rec = num_of_rec

    def to_dict(self):
        """Returns the model properties as a dict"""
        result = {}

        for attr, _ in six.iteritems(self.swagger_types):
            value = getattr(self, attr)
            if isinstance(value, list):
                result[attr] = list(map(
                    lambda x: x.to_dict() if hasattr(x, "to_dict") else x,
                    value
                ))
            elif hasattr(value, "to_dict"):
                result[attr] = value.to_dict()
            elif isinstance(value, dict):
                result[attr] = dict(map(
                    lambda item: (item[0], item[1].to_dict())
                    if hasattr(item[1], "to_dict") else item,
                    value.items()
                ))
            else:
                result[attr] = value
        if issubclass(RandomRequest, dict):
            for key, value in self.items():
                result[key] = value

        return result

    def to_str(self):
        """Returns the string representation of the model"""
        return pprint.pformat(self.to_dict())

    def __repr__(self):
        """For `print` and `pprint`"""
        return self.to_str()

    def __eq__(self, other):
        """Returns true if both objects are equal"""
        if not isinstance(other, RandomRequest):
            return False

        return self.__dict__ == other.__dict__

    def __ne__(self, other):
        """Returns true if both objects are not equal"""
        return not self == other